Add support for column_values_by_type option.
Sometimes you may have specific column values in the dump that you want
to replace with specific local values, and it's unreliable to do this
with simple string replacement. This option lets you target the broad
case with a hash, or the simple case with a lambda.

```ruby
Dumper.dump(obj, {
  dumps: [{my_active_record_type => :blog_post}],
  column_values_by_type: {
    MyActiveRecordType: {
      contact_id: 123
    },
    BlogPost: -> (blog_post) {
      { name: "#{blog_post.name} - EXPORTED FROM PROD" }
    }
  }
})
```